user_auth/requirements.txt
werkzeug==2.3.8
Flask
flask-jwt-extended
flask-cors
argon2-cffi
Flask-SQLAlchemy
secrets